fix(ci): batch -Werror cleanups for pythonpkg native/numpy/pandas/scan/wrapper
Sweep up the remaining GCC 14 -Werror diagnostics from the manylinux2014
build of integration/python (old-style-cast, sign-conversion, sign-promo,
sign-compare, reorder, unused-parameter, unused-variable, unused-value,
extra/copy-ctor base init).

In python_replacement_scan FetchObjectData also replace the stub
result_chunk.append(...) (no such method on data_chunk_t) with an explicit
chunk collection + row-by-row merge, so multi-batch pandas/numpy scans
build a single accumulated data_chunk for make_node_raw_data.

+++ integration/python/util/util.cpp
@@ -13,13 +13,13 @@ namespace otterbrix {
                 case physical_type::NA:
                     return "NULL";
                 case physical_type::BOOL:
-                    return to_string(value.value<bool>());
+                    return to_string(static_cast<unsigned int>(value.value<bool>()));
                 case physical_type::UINT8:
-                    return to_string(value.value<uint8_t>());
+                    return to_string(static_cast<unsigned int>(value.value<uint8_t>()));
                 case physical_type::INT8:
-                    return to_string(value.value<int8_t>());
+                    return to_string(static_cast<int>(value.value<int8_t>()));
                 case physical_type::UINT16:
-                    return to_string(value.value<uint16_t>());
+                    return to_string(static_cast<unsigned int>(value.value<uint16_t>()));
                 case physical_type::INT16:
                     return to_string(value.value<int16_t>());
                 case physical_type::UINT32:
